+++ module/Makefile.bsd
@@ -127,7 +127,7 @@ SRCS+= spl_atomic.c
 .endif
 
 #os/freebsd/zfs
-SRCS+=	abd.c \
+SRCS+=	abd_os.c \
 	crypto_os.c \
 	dmu_os.c \
 	hkdf.c \
@@ -169,7 +169,8 @@ SRCS+=	zfeature_common.c \
 	zprop_common.c
 
 #zfs
-SRCS+=	aggsum.c \
+SRCS+=	abd.c \
+	aggsum.c \
 	arc.c \
 	arc_os.c \
 	blkptr.c \
